+++ lib/panda_doc/failure_result.rb
@@ -1,7 +1,8 @@
 module PandaDoc
-  class FailureResult
+  class FailureResult < StandardError
     extend Forwardable
     def_delegators :response, :status, :success?
+    def_delegators :error, :type, :detail
 
     attr_reader :error
 
@@ -11,6 +12,10 @@ module PandaDoc
     def initialize(response)
       @response = response
       @error = Responses::Error.new(Objects::Error.new).from_hash(response.body)
+    end
+
+    def to_s
+      "#{status} #{type}: #{detail}"
     end
   end
 end